change float to double?

+++ variadic_functions/3-print_all.c
@@ -27,7 +27,7 @@ void print_all(const char * const format, ...)
 					printf("%s%d", sep, va_arg(mylist, int));
 					break;
 				case 'f':
-					printf("%s%f", sep, va_arg(mylist, float));
+					printf("%s%f", sep, va_arg(mylist, double));
 					break;
 				case 's':
 					str = va_arg(mylist, char *);
